backend_python/requirements.txt
fastapi
uvicorn
transformers
torch
deep-translator
langdetect
python-dotenv
openai